chore(PinterestTag): move sdk loader to different file
