+++ src/integrations/PinterestTag/browser.js
@@ -15,8 +15,8 @@ import {
   getDefinedTraits,
 } from '../../utils/utils';
 import { NAME } from './constants';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { getDestinationEventName } from './utils';
+import { loader } from './loader';
 
 export default class PinterestTag {
   constructor(config, analytics, destinationInfo) {
@@ -37,19 +37,7 @@ export default class PinterestTag {
   }
 
   loadScript() {
-    !(function (e) {
-      if (!window.pintrk) {
-        window.pintrk = function () {
-          window.pintrk.queue.push(Array.prototype.slice.call(arguments));
-        };
-        const n = window.pintrk;
-        (n.queue = []), (n.version = '3.0');
-        const t = document.createElement('script');
-        (t.async = !0), (t.src = e), t.setAttribute('data-loader', LOAD_ORIGIN);
-        const r = document.getElementsByTagName('script')[0];
-        r.parentNode.insertBefore(t, r);
-      }
-    })('https://s.pinimg.com/ct/core.js');
+    !loader('https://s.pinimg.com/ct/core.js');
   }
 
   handleEnhancedMatch() {
